# engine/docs2/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Flapjack API",
    "description": "Drop-in replacement for Algolia search. Provides full-text search with filters, facets, geo-search, and more.",
    "contact": {
      "name": "Flapjack",
      "url": "https://github.com/stuartcrobinson/flapjack"
    },
    "license": {
      "name": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:7700",
      "description": "Local development"
    },
    {
      "url": "https://fj-us-west-1.flapjack.foo",
      "description": "Production US West"
    }
  ],
  "paths": {
...
    "/1/keys": {
      "get": {
        "tags": [
          "keys"
        ],
        "summary": "List all API keys",
        "operationId": "list_keys",
        "responses": {
          "200": {
            "description": "List of keys",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListKeysResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "post": {
        "tags": [
          "keys"
        ],
        "summary": "Create a new API key with specified ACL permissions and restrictions.",
        "operationId": "create_key",
        "requestBody": {
          "description": "Key configuration",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Key created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateKeyResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateKeyRequest": {
        "type": "object",
        "description": "Deserialize an incoming API key creation request with Algolia-compatible camelCase field names.\n\nAll fields except `acl` are optional and fall back to sensible defaults (empty strings, zero, or empty vecs) when absent.",
        "required": [
          "acl"
        ],
        "properties": {
          "acl": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "indexes": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "maxHitsPerQuery": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "maxQueriesPerIPPerHour": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "queryParameters": {
            "type": [
              "string",
              "null"
            ]
          },
          "referers": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "restrictSources": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "validity": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
      "CreateKeyResponse": {
        "type": "object",
        "required": [
          "key",
          "createdAt"
        ],
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "key": {
            "type": "string"
          }
        }
      },
...
      "KeyApiResponse": {
        "type": "object",
        "description": "Algolia-compatible API response DTO for key GET/list endpoints.\nNever exposes internal fields (hash, salt, hmac_key).",
        "required": [
          "value",
          "createdAt",
          "acl",
          "description",
          "indexes",
          "maxHitsPerQuery",
          "maxQueriesPerIPPerHour",
          "queryParameters",
          "referers",
          "validity"
        ],
        "properties": {
          "acl": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "integer",
            "format": "int64",
            "description": "Epoch timestamp in milliseconds."
          },
          "description": {
            "type": "string"
          },
          "indexes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "maxHitsPerQuery": {
            "type": "integer",
            "format": "int64"
          },
          "maxQueriesPerIPPerHour": {
            "type": "integer",
            "format": "int64"
          },
          "queryParameters": {
            "type": "string"
          },
          "referers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "restrictSources": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "validity": {
            "type": "integer",
            "format": "int64"
          },
          "value": {
            "type": "string"
          }
        }
      },
...
      "ListKeysResponse": {
        "type": "object",
        "required": [
          "keys"
        ],
        "properties": {
          "keys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KeyApiResponse"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-algolia-api-key"
      }
    }
  },
...
}